new invoice PDF generation tool
The new tool is written in vala and uses cairo
and pango to generate the invoice PDF. Thus a
full-blown latex install is no longer needed.

The new tool is also much fast than generating
the invoice with latex (ca. 10x speed improvement)
and generates files, which have only half of the
size pdflatex produces.